New makefile solution: A single invocation of 'make' to build the entire tree. Fully controlled dependencies give faster and more correct recompiles.
Many #include lines adjusted to conform to the new standards.


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@19146 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/lib/codeclib.c
@@ -27,8 +27,8 @@
 #include "codeclib.h"
 #include "metadata.h"
 
-long mem_ptr;
-long bufsize;
+size_t mem_ptr;
+size_t bufsize;
 unsigned char* mp3buf;     // The actual MP3 buffer from Rockbox
 unsigned char* mallocbuf;  // 512K from the start of MP3 buffer
 unsigned char* filebuf;    // The rest of the MP3 buffer
